requirements.txt
flask
requests
python-dotenv
flask-login
flask-sqlalchemy
werkzeug
APScheduler
pdfplumber
